requirements.txt
streamlit==1.38.0
langchain==0.2.12
langchain-core==0.2.26
langchain-openai==0.1.17
langchain-qdrant==0.1.3
langgraph==0.0.57
qdrant-client==1.8.2
python-dotenv==1.0.1
pandas==2.2.2
openai==1.35.13
